feat: replace old share component with new form-based
